update the home route and refactor based on PR
